pyproject.toml
[project]
name = "yt-dbmanager"
version = "0.3.6"
description = "Add your description here"
authors = [
    { name = "TomoTom0", email = "tomoIris427+GitHub@gmail.com" }
]
dependencies = [
    "pandas>=2.2.2",
    "mysql-connector-python>=9.0.0",
    "python-dotenv>=1.0.1",
]
readme = "README.md"
requires-python = ">= 3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/yt_dbmanager"]
